Refactor token and usd decimals to utils and use in contribution added as well
